services/haystack-api/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
haystack-ai>=2.0.0
qdrant-client>=1.7.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
requests>=2.31.0
